requirements.txt
pyautogui
pyperclip
pyyaml
keyboard

